fix: only import reader/writer to decrease bundle size
We only use the reader/writer classes from protobuf.js so only
import those - drops the size of `index.min.js` by about 30kb(!).
