+++ packages/protons-runtime/src/encode.ts
@@ -1,5 +1,5 @@
 import type { Codec } from './codec.js'
-import { writer } from './writer.js'
+import { writer } from './utils.js'
 
 export function encodeMessage <T> (message: T, codec: Codec<T>): Uint8Array {
   const w = writer()
